requirements.txt
openfabric-pysdk
ollama
python-dateutil
requests
marshmallow
streamlit
pillow
chromadb
pydantic
uuid

